supervisory_agent/wrapper/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
httpx
asyncio